# Optimize DispatchProxy generated code