* Remove unnecessary allocs
* Call the Invoke method directly from generated code


### Benchmark Results

Using the following benchmark:

```C#
    [MemoryDiagnoser]
    public class DispatchProxyBenchmark
    {
        [Params(1, 100)]
        public int Count { get; set; }

        private IFoo _proxy;

        [GlobalSetup]
        public void GlobalSetup()
        {
            _proxy = CountingProxy.Wrap(new Foo());
        }

        [Benchmark]
        public void Invoke()
        {
            for (int i = 0; i < Count; i++)
            {
                _proxy.Property1 = 5;
                _proxy.Method1();
                _proxy.Method2();
            }
        }

        [Benchmark]
        public void InvokeGeneric()
        {
            for (int i = 0; i < Count; i++)
            {
                _proxy.Method3<int>(5);
                _proxy.Method3<double>(4.5);
            }
        }
    }

    interface IFoo
    {
        public int Property1 { get; set; }

        public void Method1();
        public void Method2();
        public void Method3<T>(T t);
    }

    class Foo : IFoo
    {
        public int Property1 { get; set; }

        public void Method1() { }
        public void Method2() { }
        public void Method3<T>(T t) { }
    }

    class CountingProxy : DispatchProxy
    {
        private IFoo _inner;
        public int InvocationCount { get; private set; }

        public static IFoo Wrap(IFoo inner)
        {
            IFoo wrapped = Create<IFoo, CountingProxy>();
            ((CountingProxy)wrapped)._inner = inner;
            return wrapped;
        }

        protected override object Invoke(MethodInfo targetMethod, object[] args)
        {
            InvocationCount++;

            return targetMethod.Invoke(_inner, args);
        }
    }
```

#### Before
|        Method | Count |       Mean |     Error |    StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |------ |-----------:|----------:|----------:|--------:|------:|------:|----------:|
|        Invoke |     1 |   1.134 us | 0.0256 us | 0.0295 us |  0.1156 |     - |     - |     736 B |
| InvokeGeneric |     1 |   2.276 us | 0.0370 us | 0.0309 us |  0.2273 |     - |     - |    1456 B |
|        Invoke |   100 | 111.460 us | 0.9926 us | 0.8289 us | 11.5248 |     - |     - |   73600 B |
| InvokeGeneric |   100 | 231.935 us | 6.6885 us | 7.1566 us | 22.3214 |     - |     - |  145600 B |

#### After
|        Method | Count |         Mean |       Error |      StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |------ |-------------:|------------:|------------:|--------:|------:|------:|----------:|
|        Invoke |     1 |     493.1 ns |     8.76 ns |     7.77 ns |  0.0215 |     - |     - |     136 B |
| InvokeGeneric |     1 |   1,882.9 ns |    51.62 ns |    57.37 ns |  0.1668 |     - |     - |    1056 B |
|        Invoke |   100 |  49,365.7 ns |   934.74 ns | 1,000.16 ns |  2.1351 |     - |     - |   13600 B |
| InvokeGeneric |   100 | 177,488.8 ns | 3,470.60 ns | 3,246.40 ns | 16.3352 |     - |     - |  105600 B |

Which means we are saving 200 bytes per method invocation of a DispatchProxy.

### Generated Code Comparison

In order to easily understand this change, here are the before and after of an example generated class:

#### Before

```C#

public class generatedProxy_1 : TestDispatchProxy, TypeType_GenericMethod
{
	private Action<object[]> invoke;

	public generatedProxy_1(Action<object[]> P_0)
		: this()
	{
		invoke = P_0;
	}

	public override T Echo<T>(T P_0)
	{
		object[] array = new object[1]
		{
			P_0
		};
		object[] array2 = new object[6]
		{
			this,
			typeof(TypeType_GenericMethod),
			0,
			array,
			new global::System.Type[1]
			{
				typeof(T)
			},
			null
		};
		invoke.Invoke(array2);
		return (T)array2[5];
	}
}
```

#### After

```C#
public class generatedProxy_1 : TestDispatchProxy, TypeType_GenericMethod
{
	private MethodInfo[] _methodInfos;

	public generatedProxy_1(MethodInfo[] P_0)
		: this()
	{
		_methodInfos = P_0;
	}

	public override T Echo<T>(T P_0)
	{
		object[] args = new object[1]
		{
			P_0
		};
		MethodInfo methodInfo = _methodInfos[0];
		methodInfo = methodInfo.MakeGenericMethod(typeof(T));
		object obj = ((DispatchProxy)(object)this).Invoke(methodInfo, args);
		return (T)obj;
	}
}
```