@reflectronic 2021-01-19
Are those benchmark numbers right? Is DispatchProxy.Invoke really 450 times slower now?